tools/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
elastic-transport==8.17.1
elasticsearch==9.0.1
Events==0.5
Faker==37.1.0
hurry.filesize==0.9
idna==3.10
opensearch-py==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==79.0.1
six==1.17.0
tenacity==9.1.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
